LPS-65351 Repackage classes
