+++ modules/apps/foundation/portal-security/portal-security-service-access-quota-service/src/main/java/com/liferay/portal/security/service/access/quota/internal/access/control/AccessControlPolicySAQContextFactory.java
@@ -1,0 +1,51 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.security.service.access.quota.internal.access.control;
+
+import com.liferay.portal.kernel.security.auth.AccessControlContext;
+import com.liferay.portal.security.service.access.quota.AccessControlPolicySAQMetricProvider;
+import com.liferay.portal.security.service.access.quota.internal.SAQContext;
+import com.liferay.portal.security.service.access.quota.internal.SAQContextFactory;
+import com.liferay.portal.security.service.access.quota.internal.SAQContextFactory.SAQMetricProviderAdapter;
+
+import java.lang.reflect.Method;
+
+/**
+ * @author Stian Sigvartsen
+ */
+public class AccessControlPolicySAQContextFactory
+	extends SAQContextFactory<AccessControlPolicySAQMetricProvider> {
+
+	public SAQContext buildContext(
+		final AccessControlContext accessControlContext, final Method method) {
+
+		SAQMetricProviderAdapter<AccessControlPolicySAQMetricProvider>
+			valueProvider =
+				new SAQMetricProviderAdapter
+					<AccessControlPolicySAQMetricProvider>() {
+
+			@Override
+			public String getMetricValue(
+				AccessControlPolicySAQMetricProvider saqMetricProvider) {
+					return saqMetricProvider.getMetricValue(
+						accessControlContext, method);
+			}
+
+		};
+
+		return super.buildSaqContext(valueProvider);
+	}
+
+}
